Rename 4a and 4b tests
